requirements.txt
python-dotenv
openai
google-generativeai
groq
pymongo
fastapi
uvicorn
pydantic-settings
